server/requirements.txt
# Dependencies for the Legominist FastAPI server
pymupdf==1.25.3
chromadb==1.5.5
google-genai
python-dotenv==1.1.0
rich==14.0.0
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.11.3
